Reducing number of CodePro audit warnings

+++ trunk/yaji-ecmascript-interpreter/src/FESI/ClassFile/ConstantPoolEntry.java
@@ -95,7 +95,7 @@ abstract class ConstantPoolEntry {
         if (index == -1)
             index = classFile.addConstantPoolEntry(this);
 
-        return (short) index;
+        return index;
     }
 
     /**
